Split model tests, fix create_output graph bug
